pyproject.toml
[project]
name = "weather-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.1.0",
    "openai-agents>=0.10.5",
    "python-dotenv>=1.2.2",
]
